tested the matching implementation of SIFT_GPU now it works without external functions. added conversion function from pos, desc to GPU keypoints meaning we can now run gpu matching with different features, not only SIFT_GPU features. Tested Import Error of Silx.
