+++ opensfm/feature_loading.py
@@ -10,7 +10,6 @@ from repoze.lru import LRUCache
 
 from opensfm import features as ft
 
-
 logger = logging.getLogger(__name__)
 
 
@@ -104,12 +103,11 @@ class FeatureLoader(object):
             points = np.array(points[:, :3], dtype=float)
         return points, features, colors
 
-
     def create_gpu_keypoints_from_features(self, p1, f1):
         ########################################################################
         # Merge keypoints in central memory
         ########################################################################
-        if np.dtype(f1) is not np.uint8:
+        if type(f1[0][0]) is not int:
             f1 = np.uint8(f1 * 512)
         total_size = len(p1)
         dtype_kp = np.dtype([('x', np.float32),
@@ -119,14 +117,9 @@ class FeatureLoader(object):
                              ('desc', (np.uint8, 128))
                              ])
         output = np.recarray(shape=(total_size,), dtype=dtype_kp)
-        last = 0
-        for ds, desc in zip(p1, f1):
-            l = ds.shape[0]
-            if l > 0:
-                output[last:last + l].x = ds[:, 0]
-                output[last:last + l].y = ds[:, 1]
-                output[last:last + l].scale = ds[:, 2]
-                output[last:last + l].angle = ds[:, 3]
-                output[last:last + l].desc = desc
-                last += l
+        output[:total_size].x = p1[:, 0]
+        output[:total_size].y = p1[:, 1]
+        output[:total_size].scale = p1[:, 2]
+        # output[:total_size].angle = p1[:, 3]
+        output[:total_size].desc = f1
         return output
