+++ opensfm/matching.py
@@ -12,9 +12,18 @@ from opensfm import log
 from opensfm import multiview
 from opensfm import pairs_selection
 from opensfm import feature_loader
-from opensfm.sift_gpu import SiftGpu, check_gpu_initialization
+from opensfm.sift_gpu import SiftGpu
 
 logger = logging.getLogger(__name__)
+
+
+def check_gpu_initialization(config, image, data=None):
+    if 'sift_gpu' not in globals():
+        global sift_gpu
+        if data is not None:
+            sift_gpu = SiftGpu.sift_gpu_from_config(config, data.load_image(image))
+        else:
+            sift_gpu = SiftGpu.sift_gpu_from_config(config, image)
 
 
 def clear_cache():
@@ -40,8 +49,6 @@ def match_images(data, ref_images, cand_images):
         ref_images, cand_images, exifs, data)
 
     # Match them !
-    if config["feature_type"] == "SIFT_GPU":
-        return match_images_with_pairs_gpu(data, exifs, ref_images, pairs), preport
     return match_images_with_pairs(data, exifs, ref_images, pairs), preport
 
 
@@ -63,9 +70,6 @@ def match_images_with_pairs(data, exifs, ref_images, pairs):
     logger.info('Matching {} image pairs'.format(len(pairs)))
     mem_per_process = 512
     jobs_per_process = 2
-    if data.config['feature_type'] is 'SIFT_GPU':
-        data.config['processes'] = 1  # GPU cant run in parallel but it faster even when it does not run in parallel
-        check_gpu_initialization(data.config, pairs[0][0], data=data)
     processes = context.processes_that_fit_in_memory(data.config['processes'], mem_per_process)
     logger.info("Computing pair matching with %d processes" % processes)
     matches = context.parallel_map(match_unwrap_args, args, processes, jobs_per_process)
@@ -84,107 +88,6 @@ def match_images_with_pairs(data, exifs, ref_images, pairs):
         for im2, m in im1_matches.items():
             resulting_pairs[im1, im2] = m
 
-    return resulting_pairs
-
-
-def match_all_images_gpu(data, exifs, per_images):
-    cameras = data.load_camera_models()
-    pairs = sorted(per_images.items(), key=lambda x: -len(x[1]))
-    matches = []
-    for img1, candidates in pairs:
-        img1_matches = {}
-        camera1 = cameras[exifs[img1]['camera']]
-        for img2 in candidates:
-            camera2 = cameras[exifs[img2]['camera']]
-            img1_matches[img2] = match_images_sift_gpu(data, img1, img2, camera1, camera2)
-        matches.append([img1, img1_matches])
-    return matches
-
-
-def match_images_sift_gpu(data, img1, img2, camera1, camera2):
-    config = data.config
-    time_start = timer()
-    p1, f1, _ = feature_loader.instance.load_points_features_colors(
-        data, img1, masked=False)
-    p2, f2, _ = feature_loader.instance.load_points_features_colors(
-        data, img2, masked=False)
-    keypoints1 = data.load_gpu_features(img1)
-    keypoints2 = data.load_gpu_features(img2)
-
-    if p1 is None or len(p1) < 2 or p2 is None or len(p2) < 2:
-        return []
-    matches = sift_gpu.match_images(keypoints1, keypoints2)
-
-    # Adhoc filters
-    if config['matching_use_filters']:
-        matches = apply_adhoc_filters(data, matches,
-                                      img1, camera1, p1,
-                                      img2, camera2, p2)
-    matches = np.array(matches, dtype=int)
-    time_2d_matching = timer() - time_start
-    t = timer()
-    robust_matching_min_match = config['robust_matching_min_match']
-    if len(matches) < robust_matching_min_match:
-        logger.debug(
-            'Matching {} and {}.  Matcher: {} T-desc: {:1.3f} '
-            'Matches: FAILED'.format(
-                img1, img2,
-                "SIFT_GPU",
-                time_2d_matching))
-        return []
-
-    # robust matching
-    rmatches = robust_match(p1, p2, camera1, camera2, matches, config)
-    rmatches = np.array([[a, b] for a, b in rmatches])
-    time_robust_matching = timer() - t
-    time_total = timer() - time_start
-
-    # From indexes in filtered sets, to indexes in original sets of features
-    m1 = feature_loader.instance.load_mask(data, img1)
-    m2 = feature_loader.instance.load_mask(data, img2)
-    if m1 is not None and m2 is not None:
-        rmatches = unfilter_matches(rmatches, m1, m2)
-
-    logger.debug(
-        'Matching {} and {}.  Matcher: {}  '
-        'T-desc: {:1.3f} T-robust: {:1.3f} T-total: {:1.3f} '
-        'Matches: {} Robust: {} Success: {}'.format(
-            img1, img2, "SIFT_GPU",
-            time_2d_matching, time_robust_matching, time_total,
-            len(matches), len(rmatches),
-            len(rmatches) >= robust_matching_min_match))
-
-    if len(rmatches) < robust_matching_min_match:
-        return []
-
-    return np.array(rmatches, dtype=int)
-
-
-def match_images_with_pairs_gpu(data, exifs, ref_images, pairs):
-    """ Perform pair matchings given pairs. """
-
-    # Store per each image in ref for processing
-    per_image = {im: [] for im in ref_images}
-    check_gpu_initialization(data.config, data.load_image(pairs[0][0]))
-    for im1, im2 in pairs:
-        per_image[im1].append(im2)
-
-    start = timer()
-    matches = match_all_images_gpu(data, exifs, per_image)
-
-    logger.info(
-        'Matched {} pairs for {} ref_images '
-        'in {} seconds ({} seconds/pair).'.format(
-            len(pairs),
-            len(ref_images),
-            timer() - start,
-            (timer() - start) / len(pairs) if pairs else 0))
-
-    # Index results per pair
-    resulting_pairs = {}
-    for im1, im1_matches in matches:
-        for im2, m in im1_matches.items():
-            resulting_pairs[im1, im2] = m
     return resulting_pairs
 
 
@@ -299,11 +202,12 @@ def match(im1, im2, camera1, camera2, data):
         else:
             matches = match_brute_force(f1, f2, config)
     elif matcher_type == 'SIFT_GPU':
+        check_gpu_initialization(config, im1, data)
         k1 = data.load_gpu_features(im1)
         k2 = data.load_gpu_features(im2)
         if k1 is None or k2 is None:
-            k1 = feature_loader.instance.create_gpu_keypoints_from_features(f1, p1)
-            k2 = feature_loader.instance.create_gpu_keypoints_from_features(f2, p2)
+            k1 = feature_loader.instance.create_gpu_keypoints_from_features(p1, f1)
+            k2 = feature_loader.instance.create_gpu_keypoints_from_features(p2, f2)
         matches = sift_gpu.match_images(k1, k2)
     else:
         raise ValueError("Invalid matcher_type: {}".format(matcher_type))
